Refactor
  * Bugfix in `negotiateMediaTypeFactory` regarding `permissive` flag
  * Rename `parseAcceptHeader` to `rawParseAcceptHeader`
  * New `parseAcceptHeader` function that makes it more suitable for
    standalone use.
  * Better compatibility with `negotiator`
