+++ src/negotiateMediaTypeFactory.ts
@@ -14,44 +14,11 @@
  */
 
 import { $subtype, $type } from './constants.js';
+import findQ from './lib/findQ.js';
 import normaliseMediaType from './normaliseMediaType.js';
-import parseAcceptHeader from './parseAcceptHeader.js';
+import rawParseAcceptHeader from './rawParseAcceptHeader.js';
 import parseMediaType, { type TMediaType } from './parseMediaType.js';
 import { wm } from './utils.js';
-
-const QVALUE_REGEX = /^(?:(?:0(?:\.\d{1,3})?)|(?:1(?:\.0{1,3})?))$/;
-const DEFAULT_Q = 1000;
-const MAX_Q = 1000;
-const MIN_Q = 0;
-
-/**
- * Convert a q-value (quality factor) string to a clamped integer weight
- * (0–1000).
- *
- * Q-values follow RFC semantics (0 to 1, up to three decimal places). This
- * module uses integer weights in the range **0..1000** where `1.0` maps to
- * **1000** and `0.5` maps to **500**, &c. Invalid or missing `q` parameters
- * default to **1000**.
- *
- * @internal
- * @param mediaType - Normalised media type tuple (parameters available
- *   at index 1).
- * @returns Integer q weight in the range **0..1000**.
- */
-const findQ = (mediaType: TMediaType): number => {
-	const qParam = mediaType[1].find((p) => p[0] === 'q');
-	if (!qParam || !QVALUE_REGEX.test(qParam[1])) return DEFAULT_Q;
-
-	// "0.5" -> "05" -> "0500", "1" -> "1" -> "1000"
-	const digits = qParam[1].replace('.', '');
-	const normalised = digits.padEnd(4, '0');
-	const qvalue = parseInt(normalised, 10);
-	// NaN check
-	if (qvalue !== qvalue) return DEFAULT_Q;
-
-	// Clamp in case something failed
-	return Math.max(Math.min(MAX_Q, qvalue), MIN_Q);
-};
 
 /**
  * Find parameters that overlap between two media types, excluding the `q`
@@ -139,11 +106,7 @@ const negotiateMediaTypeFactory = (availableMediaTypes: string[]) => {
 
 		const qMap = wm<TMediaType, number>();
 
-		const parsedAcceptableTypes = parseAcceptHeader(
-			accept,
-			permissive,
-			false,
-		)
+		const parsedAcceptableTypes = rawParseAcceptHeader(accept, permissive)
 			.map((type) => {
 				const parsed = parseMediaType(type, permissive);
 				const normalised = normaliseMediaType(parsed);
