+++ src/parseAcceptHeader.ts
@@ -13,272 +13,122 @@
  * PERFORMANCE OF THIS SOFTWARE.
  */
 
-import { isOWS, isToken } from './utils.js';
-
-const STATE_INVALID = -1 as const;
-const STATE_INITIAL = 0 as const;
-const STATE_TYPE = 1 as const;
-const STATE_SUBTYPE_START = 2 as const;
-const STATE_SUBTYPE = 3 as const;
-const STATE_SUBTYPE_END = 4 as const;
-const STATE_PARAMS_START = 5 as const;
-const STATE_PARAMETER_NAME = 6 as const;
-const STATE_PARAMETER_NAME_END = 7 as const;
-const STATE_PARAMETER_VALUE_START = 8 as const;
-const STATE_PARAMETER_VALUE = 9 as const;
-const STATE_PARAMETER_QUOTED = 10 as const;
-
-type TState =
-	| typeof STATE_INVALID
-	| typeof STATE_INITIAL
-	| typeof STATE_TYPE
-	| typeof STATE_SUBTYPE_START
-	| typeof STATE_SUBTYPE
-	| typeof STATE_SUBTYPE_END
-	| typeof STATE_PARAMS_START
-	| typeof STATE_PARAMETER_NAME
-	| typeof STATE_PARAMETER_NAME_END
-	| typeof STATE_PARAMETER_VALUE_START
-	| typeof STATE_PARAMETER_VALUE
-	| typeof STATE_PARAMETER_QUOTED;
+import findQ from './lib/findQ.js';
+import type { TMediaType } from './parseMediaType.js';
+import parseMediaType from './parseMediaType.js';
+import rawParseAcceptHeader from './rawParseAcceptHeader.js';
+import { wm } from './utils.js';
 
 /**
- * Parse an HTTP Accept header value into an array of media-range strings.
+ * Parse an HTTP Accept header into an array of media type strings, with
+ * optional filtering, sorting by quality factors (q), and the option to return
+ * only type/subtype.
  *
- * The parser implements a small state machine mirroring RFC 9110 media range
- * syntax and is intentionally lightweight: it extracts each "type/subtype"
- * media-range (including any parameters when `typesOnly` is false) as raw
- * substrings from the input without allocating objects for parameters.
+ * This function wraps a lower-level parser (`rawParseAcceptHeader`) that splits
+ * the header into raw media type strings, then optionally parses each media
+ * type into its structured representation (`parseMediaType`) to inspect `q`
+ * values and other parameters. It supports permissive parsing mode for
+ * non-strict inputs.
  *
- * Features:
- * - Recognises `token` and `OWS` via helpers.
- * - Supports permissive mode to tolerate common non-RFC inputs (extra OWS,
- *   empty parameter values, flag parameters, and truncated quoted values at
- *   EOF).
- * - When `typesOnly` is `true`, only the bare "type/subtype" substrings are
- *   returned; when `false` (default), the returned strings include trailing
- *   parameters as they appear in the header.
+ * Behaviour summary:
+ * - If `accept` is empty or undefined, returns `['*' + '/*']`.
+ * - If neither `typesOnly` nor `sort` are requested, returns that raw list
+ *   immediately (fast path).
+ * - Otherwise, parses each media-type string into a `TMediaType` structure.
+ * - Filters out media types with a zero `q` value.
+ * - If `sort` is true, sorts the remaining types in descending order by `q`.
+ * - If `typesOnly` is true, returns just the `type/subtype` strings
+ *   (e.g. "text/html").
+ *   If `typesOnly` is false, returns the original raw media-type strings
+ *   (including parameters).
  *
- * Behaviour details:
- * - Returns an array of strings in the order they appear in the header.
- * - Does not perform case-normalisation; callers should normalise if needed.
- * - In strict mode (`permissive=false`), malformed segments cause them to be
- *   skipped or the parser to enter an invalid state; parsing continues at the
- *   next comma boundary.
+ * Notes:
+ * - The function uses weak maps internally to associate parsed structures with
+ *   original strings and to store q-values for sorting; these are not exposed.
+ * - `permissive` is forwarded to the underlying parsers to allow tolerant
+ *    parsing of malformed headers.
  *
- * @param accept - The raw Accept header value to parse.
- * @param typesOnly - If true, return only the "type/subtype" portion of each
- *   media-range. If false, include parameters (e.g. "text/plain; q=0.5") in the
- *   returned strings.
- * @param permissive - If true, accept some non-RFC-compliant inputs (extra OWS,
- *   empty parameter values, flag parameters like `;foo`, and unterminated
- *   quoted values at EOF).
- * @returns {string[]} Array of media-range strings (either types only or with
- *   parameters), in the same order they appeared in the input.
+ * Example:
+ *   parseAcceptHeader('text/html;q=0.8, application/json;q=0.9', true)
+ *   // => ['application/json;q=0.9', 'text/html;q=0.8']
  *
- * @example
- * parseAcceptHeader('text/html, text/plain;q=0.8, application/json');
- * // -> ['text/html', 'text/plain;q=0.8', 'application/json']
+ * Example (typesOnly):
+ *   parseAcceptHeader('text/html;q=0.8, application/json;q=0.9', true, true)
+ *   // => ['application/json', 'text/html']
  *
- * @example
- * parseAcceptHeader('text/*;q=0.5, text/plain', true);
- * // -> ['text/*', 'text/plain']
+ * @param accept The raw value of an HTTP Accept header. If omitted or falsy,
+ * the function returns ['*' + '/*'] as the default.
+ *
+ * @param sort If true, returned media types are sorted in descending order of
+ * their quality parameter. If false, the original order is preserved
+ * (unless filtering removes entries).
+ *
+ * @param typesOnly If true, the function returns only the "type/subtype"
+ * portion of each media type (e.g., 'text/html'), omitting any parameters.
+ * If false, original raw media-type strings (with parameters) are returned.
+ *
+ * @param permissive Passed to the underlying parsers to enable permissive
+ * tolerances when parsing malformed or non-standard media type strings. When
+ * true, the parsers attempts to salvage non-RFC complient inputs.
+ *
+ * @returns An array of media-type strings. When `typesOnly` is true, elements
+ * are `type/subtype` strings; otherwise they are the original media-type
+ * strings provided by the client. If `accept` is falsy, returns ['*' + '/*'].
  */
 const parseAcceptHeader = (
-	accept: string,
+	accept?: string,
+	sort?: boolean,
 	typesOnly?: boolean,
 	permissive?: boolean,
 ): string[] => {
-	let state: TState = STATE_INITIAL;
-	const types: string[] = [];
-
-	let pos = 0;
-	let s = 0,
-		te = 0,
-		pe = 0;
-
-	const reset = () => {
-		state = STATE_INITIAL;
-		types.push(accept.slice(s, typesOnly ? te : pe));
-	};
-
-	for (; pos <= accept.length; pos++) {
-		const chr = accept[pos];
-		switch (state) {
-			case STATE_INITIAL: {
-				if (isOWS(chr) || chr === ',') {
-					continue;
-				} else if (isToken(chr)) {
-					state = STATE_TYPE;
-					s = pos;
-				} else {
-					state = STATE_INVALID;
-				}
-				break;
-			}
-			case STATE_TYPE: {
-				if (isToken(chr)) {
-					continue;
-				} else if (chr === '/') {
-					state = STATE_SUBTYPE_START;
-				} else if (chr === ',') {
-					// reset and continue
-					state = STATE_INITIAL;
-				} else {
-					state = STATE_INVALID;
-				}
-				break;
-			}
-			case STATE_SUBTYPE_START: {
-				if (isToken(chr)) {
-					state = STATE_SUBTYPE;
-				} else if (permissive && isOWS(chr)) {
-					continue;
-				} else if (chr === ',') {
-					// reset and continue
-					state = STATE_INITIAL;
-				} else {
-					state = STATE_INVALID;
-				}
-				break;
-			}
-			case STATE_SUBTYPE: {
-				if (isToken(chr)) {
-					continue;
-				} else if (chr === ',' || chr === undefined) {
-					te = pe = pos;
-					reset();
-				} else if (chr === ';') {
-					te = pe = pos;
-					state = STATE_PARAMS_START;
-				} else if (isOWS(chr)) {
-					te = pe = pos;
-					state = STATE_SUBTYPE_END;
-				} else {
-					state = STATE_INVALID;
-				}
-				break;
-			}
-			case STATE_SUBTYPE_END: {
-				if (chr === ',' || chr === undefined) {
-					reset();
-				} else if (chr === ';') {
-					state = STATE_PARAMS_START;
-				} else if (isOWS(chr)) {
-					continue;
-				} else {
-					state = STATE_INVALID;
-				}
-				break;
-			}
-			case STATE_PARAMS_START: {
-				if (isToken(chr)) {
-					state = STATE_PARAMETER_NAME;
-				} else if (isOWS(chr) || chr === ';') {
-					continue;
-				} else if (chr === ',' || chr === undefined) {
-					reset();
-				} else {
-					state = STATE_INVALID;
-				}
-				break;
-			}
-			case STATE_PARAMETER_NAME: {
-				if (isToken(chr)) {
-					continue;
-				} else if (chr === '=') {
-					state = STATE_PARAMETER_VALUE_START;
-				} else if (permissive && isOWS(chr)) {
-					pe = pos;
-					state = STATE_PARAMETER_NAME_END;
-				} else if (permissive && chr === ';') {
-					// Laxer than the RFC mandates
-					// 'Flag' parameter, like `example/example; foo; bar=baz`
-					state = STATE_PARAMS_START;
-				} else if (permissive && (chr === ',' || chr === undefined)) {
-					reset();
-				} else {
-					state = STATE_INVALID;
-				}
-				break;
-			}
-			case STATE_PARAMETER_NAME_END: {
-				if (isOWS(chr)) {
-					continue;
-				} else if (chr === ';') {
-					state = STATE_PARAMS_START;
-				} else if (chr === '=') {
-					if (permissive) {
-						pe = pos + 1;
-					}
-					state = STATE_PARAMETER_VALUE_START;
-				} else if (chr === ',' || chr === undefined) {
-					reset();
-				}
-				break;
-			}
-			case STATE_PARAMETER_VALUE_START: {
-				if (isToken(chr)) {
-					state = STATE_PARAMETER_VALUE;
-				} else if (chr === '"') {
-					state = STATE_PARAMETER_QUOTED;
-				} else if (permissive && isOWS(chr)) {
-					// Laxer than the RFC mandates
-					continue;
-				} else if (permissive && chr === ';') {
-					// Laxer than the RFC mandates, as empty parameter values
-					// allowed
-					state = STATE_PARAMS_START;
-				} else if (permissive && (chr === ',' || chr === undefined)) {
-					// Laxer than the RFC mandates, as empty parameter values
-					// allowed
-					pe = pos;
-					reset();
-				} else {
-					state = STATE_INVALID;
-				}
-				break;
-			}
-			case STATE_PARAMETER_VALUE: {
-				if (isToken(chr)) {
-					continue;
-				} else if (chr === ';') {
-					pe = pos;
-					state = STATE_PARAMS_START;
-				} else if (chr === ',' || chr === undefined) {
-					pe = pos;
-					reset();
-				} else if (isOWS(chr)) {
-					pe = pos;
-					state = STATE_SUBTYPE_END;
-				} else {
-					state = STATE_INVALID;
-				}
-				break;
-			}
-			case STATE_PARAMETER_QUOTED: {
-				if (permissive && chr === undefined) {
-					pe = pos;
-					reset();
-				} else if (chr !== '"' || accept[pos - 1] === '\\') {
-					continue;
-				} else {
-					state = STATE_SUBTYPE_END;
-					pe = pos + 1;
-				}
-				break;
-			}
-			case STATE_INVALID: {
-				if (chr === ',') {
-					// reset and continue
-					state = STATE_INITIAL;
-				}
-				break;
-			}
-		}
+	if (!accept) {
+		return ['*/*'];
 	}
 
-	return types;
+	const result = rawParseAcceptHeader(accept, permissive);
+
+	if (!typesOnly && !sort) {
+		return result;
+	}
+
+	const qmap = wm<TMediaType, number>();
+	const dmap = typesOnly ? undefined : wm<TMediaType, string>();
+	const parsed = result.map((t) => {
+		const p = parseMediaType(t, permissive);
+		if (!typesOnly) {
+			dmap!.set(p, t);
+		}
+		return p;
+	});
+
+	const filtered = parsed.filter((t) => {
+		const q = findQ(t);
+		if (!q) return false;
+
+		qmap.set(t, q);
+		return true;
+	});
+
+	if (sort) {
+		filtered.sort((a, b) => {
+			const qa = qmap.get(a)!;
+			const qb = qmap.get(b)!;
+
+			return qb - qa;
+		});
+	}
+
+	const list = filtered.map(
+		typesOnly
+			? (t) => {
+					return t[0];
+				}
+			: (t) => {
+					return dmap!.get(t)!;
+				},
+	);
+
+	return list;
 };
 
 export default parseAcceptHeader;
