+++ test/rawParseAcceptHeader.test.ts
@@ -15,16 +15,16 @@
 
 import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
-import parseAcceptHeader from '../src/parseAcceptHeader.js';
+import rawParseAcceptHeader from '../src/rawParseAcceptHeader.js';
 
-describe('parseAcceptHeader', () => {
+describe('rawParseAcceptHeader', () => {
 	// Basic well-formed headers
 	it('parses simple single type', () => {
-		assert.deepEqual(parseAcceptHeader('text/html'), ['text/html']);
+		assert.deepEqual(rawParseAcceptHeader('text/html'), ['text/html']);
 	});
 
 	it('parses multiple types separated by comma', () => {
-		assert.deepEqual(parseAcceptHeader('text/html, application/json'), [
+		assert.deepEqual(rawParseAcceptHeader('text/html, application/json'), [
 			'text/html',
 			'application/json',
 		]);
@@ -32,51 +32,47 @@ describe('parseAcceptHeader', () => {
 
 	it('preserves order of appearance', () => {
 		assert.deepEqual(
-			parseAcceptHeader('application/xml, text/plain, image/png'),
+			rawParseAcceptHeader('application/xml, text/plain, image/png'),
 			['application/xml', 'text/plain', 'image/png'],
 		);
 	});
 
 	// Wildcards
 	it('parses wildcard types', () => {
-		assert.deepEqual(parseAcceptHeader('*/*'), ['*/*']);
-		assert.deepEqual(parseAcceptHeader('image/*, */*'), ['image/*', '*/*']);
+		assert.deepEqual(rawParseAcceptHeader('*/*'), ['*/*']);
+		assert.deepEqual(rawParseAcceptHeader('image/*, */*'), [
+			'image/*',
+			'*/*',
+		]);
 	});
 
 	// Parameters (q, charset, etc.)
 	it('preserves parameters', () => {
-		assert.deepEqual(parseAcceptHeader('text/html; charset=UTF-8; q=0.9'), [
-			'text/html; charset=UTF-8; q=0.9',
-		]);
-	});
-
-	it('ignores parameters and returns only type/subtype', () => {
 		assert.deepEqual(
-			parseAcceptHeader(
-				'text/html; charset=UTF-8; q=0.9, example/plain, application/example ; q=0.4, test/test',
-				true,
-			),
-			['text/html', 'example/plain', 'application/example', 'test/test'],
+			rawParseAcceptHeader('text/html; charset=UTF-8; q=0.9'),
+			['text/html; charset=UTF-8; q=0.9'],
 		);
 	});
 
 	it('parses multiple types with parameters', () => {
 		assert.deepEqual(
-			parseAcceptHeader('text/html; q=0.8, application/json; q=0.9'),
+			rawParseAcceptHeader('text/html; q=0.8, application/json; q=0.9'),
 			['text/html; q=0.8', 'application/json; q=0.9'],
 		);
 	});
 
 	it('handles quoted parameter values containing commas and semicolons', () => {
 		assert.deepEqual(
-			parseAcceptHeader('text/plain; title="a, b; c", application/json'),
+			rawParseAcceptHeader(
+				'text/plain; title="a, b; c", application/json',
+			),
 			['text/plain; title="a, b; c"', 'application/json'],
 		);
 	});
 
 	it('handles escaped quotes inside quoted parameter values', () => {
 		assert.deepEqual(
-			parseAcceptHeader(
+			rawParseAcceptHeader(
 				'text/plain; title="a \\"quoted\\" text"; q=0.5, image/png',
 			),
 			['text/plain; title="a \\"quoted\\" text"; q=0.5', 'image/png'],
@@ -86,7 +82,7 @@ describe('parseAcceptHeader', () => {
 	// OWS and unusual whitespace
 	it('handles optional whitespace around tokens and separators', () => {
 		assert.deepEqual(
-			parseAcceptHeader(
+			rawParseAcceptHeader(
 				'  text/html  ; q=0.7 ,application/json\t; q=0.8 ',
 			),
 			['text/html  ; q=0.7', 'application/json\t; q=0.8'],
@@ -95,29 +91,29 @@ describe('parseAcceptHeader', () => {
 
 	// Empty and whitespace-only input
 	it('returns empty array for empty string', () => {
-		assert.deepEqual(parseAcceptHeader(''), []);
+		assert.deepEqual(rawParseAcceptHeader(''), []);
 	});
 
 	it('returns empty array for whitespace-only string', () => {
-		assert.deepEqual(parseAcceptHeader('   \t  '), []);
+		assert.deepEqual(rawParseAcceptHeader('   \t  '), []);
 	});
 
 	// Malformed inputs (strict mode default)
 	it('skips malformed entries but continues with later valid ones', () => {
 		assert.deepEqual(
-			parseAcceptHeader('application/json, bad@@type, text/plain'),
+			rawParseAcceptHeader('application/json, bad@@type, text/plain'),
 			['application/json', 'text/plain'],
 		);
 	});
 
 	it('ignores entries missing subtype', () => {
-		assert.deepEqual(parseAcceptHeader('text/, application/json'), [
+		assert.deepEqual(rawParseAcceptHeader('text/, application/json'), [
 			'application/json',
 		]);
 	});
 
 	it('ignores stray slashes and extra commas', () => {
-		assert.deepEqual(parseAcceptHeader(',,/, , application/xml, ,'), [
+		assert.deepEqual(rawParseAcceptHeader(',,/, , application/xml, ,'), [
 			'application/xml',
 		]);
 	});
@@ -125,7 +121,7 @@ describe('parseAcceptHeader', () => {
 	// Permissive mode behaviour (allows some lax constructs)
 	it('in permissive mode accepts flag parameters and empty param values', () => {
 		const input = 'example/example; foo; bar=; q=0.5, text/plain';
-		assert.deepEqual(parseAcceptHeader(input, false, true), [
+		assert.deepEqual(rawParseAcceptHeader(input, true), [
 			'example/example; foo; bar=; q=0.5',
 			'text/plain',
 		]);
@@ -133,7 +129,7 @@ describe('parseAcceptHeader', () => {
 
 	it('in permissive mode tolerates OWS inside parameter handling', () => {
 		const input = 'text/plain;   charset = utf-8  , application/json';
-		assert.deepEqual(parseAcceptHeader(input, false, true), [
+		assert.deepEqual(rawParseAcceptHeader(input, true), [
 			'text/plain;   charset = utf-8',
 			'application/json',
 		]);
@@ -141,7 +137,7 @@ describe('parseAcceptHeader', () => {
 
 	it('in permissive mode tolerates EOF inside quoted string', () => {
 		const input = 'text/plain; title="unterminated';
-		assert.deepEqual(parseAcceptHeader(input, false, true), [
+		assert.deepEqual(rawParseAcceptHeader(input, true), [
 			'text/plain; title="unterminated',
 		]);
 	});
@@ -152,12 +148,12 @@ describe('parseAcceptHeader', () => {
 			{ length: 100 },
 			(_, i) => `type${i}/sub${i}`,
 		).join(',');
-		assert.deepEqual(parseAcceptHeader(many).length, 100);
+		assert.deepEqual(rawParseAcceptHeader(many).length, 100);
 	});
 
 	it('retains case for returned media-ranges', () => {
 		assert.deepEqual(
-			parseAcceptHeader('TEXT/HTML; Q=0.8, Application/JSON'),
+			rawParseAcceptHeader('TEXT/HTML; Q=0.8, Application/JSON'),
 			['TEXT/HTML; Q=0.8', 'Application/JSON'],
 		);
 	});
@@ -165,7 +161,7 @@ describe('parseAcceptHeader', () => {
 	// Entries with q values but ordering preserved (no sorting by q)
 	it('does not sort by q; preserves original order', () => {
 		const input = 'text/plain;q=0.1, text/html;q=1.0';
-		assert.deepEqual(parseAcceptHeader(input), [
+		assert.deepEqual(rawParseAcceptHeader(input), [
 			'text/plain;q=0.1',
 			'text/html;q=1.0',
 		]);
@@ -174,7 +170,7 @@ describe('parseAcceptHeader', () => {
 	// Ensure quoted param with escaped backslash then quote handled
 	it('handles escaped backslash before quote inside quoted parameter', () => {
 		const input = 'text/plain; title="one\\\\"two" , image/jpeg';
-		assert.deepEqual(parseAcceptHeader(input), [
+		assert.deepEqual(rawParseAcceptHeader(input), [
 			'text/plain; title="one\\\\"two"',
 			'image/jpeg',
 		]);
