+++ tsconfig.json
@@ -21,6 +21,6 @@
 	"ts-node": {
 		"experimentalResolver": true
 	},
-	"include": ["src/**/*"],
-	"exclude": ["node_modules", "**/*.spec.ts", "**/*.test.ts"]
+	"include": ["src/**/*", "@types/*.d.ts"],
+	"exclude": ["node_modules"]
 }
